factor out an indexFile base class from bai.js and csi.js
