+++ src/bai.js
@@ -1,30 +1,12 @@
 const Long = require('long')
 const VirtualOffset = require('./virtualOffset')
 const Chunk = require('./chunk')
+const IndexFile = require('./indexFile')
 
 const BAI_MAGIC = 21578050 // BAI\1
 const { longToNumber, abortBreakPoint } = require('./util')
 
-class BAI {
-  /**
-   * @param {filehandle} filehandle
-   * @param {function} [renameRefSeqs]
-   */
-  constructor({ filehandle, renameRefSeqs = n => n }) {
-    this.filehandle = filehandle
-    this.renameRefSeq = renameRefSeqs
-  }
-
-  _findFirstData(data, virtualOffset) {
-    const currentFdl = data.firstDataLine
-    if (currentFdl) {
-      data.firstDataLine =
-        currentFdl.compareTo(virtualOffset) > 0 ? virtualOffset : currentFdl
-    } else {
-      data.firstDataLine = virtualOffset
-    }
-  }
-
+class BAI extends IndexFile {
   parsePseudoBin(bytes, offset) {
     const lineCount = longToNumber(
       Long.fromBytesLE(bytes.slice(offset + 16, offset + 24), true),
@@ -39,22 +21,6 @@ class BAI {
     }
     const ret = index.stats || {}
     return ret.lineCount === undefined ? -1 : ret.lineCount
-  }
-
-  async parse(abortSignal) {
-    if (!this._parseCache) {
-      this._parseCache = this._parse(abortSignal)
-      this._parseCache.catch(() => {
-        if (abortSignal && abortSignal.aborted) delete this._parseCache
-      })
-      return this._parseCache
-    }
-    return this._parseCache.catch(e => {
-      if (e.code === 'ERR_ABORTED' || e.name === 'AbortError') {
-        return this.parse(abortSignal)
-      }
-      throw e
-    })
   }
 
   // fetch and parse the index
@@ -194,16 +160,6 @@ class BAI {
     numOffsets = l + 1
 
     return off.slice(0, numOffsets)
-  }
-
-  /**
-   * @param {number} seqId
-   * @param {AbortSignal} [abortSignal]
-   * @returns {Promise} true if the index contains entries for
-   * the given reference sequence ID, false otherwise
-   */
-  async hasRefSeq(seqId, abortSignal) {
-    return !!((await this.parse(abortSignal)).indices[seqId] || {}).binIndex
   }
 
   /**
